+++ aws_lambda/alexa_smart_home/__init__.py
@@ -1,0 +1,1 @@
+"""Alexa Smart Home Lambda adapter."""